test: Replace slider part test with testing lib
